+++ protocol-units/sequencing/memseq/sequencer/src/lib.rs
@@ -1,6 +1,9 @@
 use mempool_util::{MempoolBlockOperations, MempoolTransactionOperations};
 pub use move_rocks::RocksdbMempool;
-pub use movement_types::{Block, Id, Transaction};
+pub use movement_types::{
+	block::{self, Block},
+	transaction::{self, Transaction},
+};
 pub use sequencing_util::Sequencer;
 use std::collections::BTreeSet;
 use std::{path::PathBuf, sync::Arc};
@@ -11,7 +14,7 @@ pub struct Memseq<T: MempoolBlockOperations + MempoolTransactionOperations> {
 	mempool: T,
 	// this value should not be changed after initialization
 	block_size: u32,
-	pub parent_block: Arc<RwLock<Id>>,
+	pub parent_block: Arc<RwLock<block::Id>>,
 	// this value should not be changed after initialization
 	building_time_ms: u64,
 }
@@ -20,7 +23,7 @@ impl<T: MempoolBlockOperations + MempoolTransactionOperations> Memseq<T> {
 	pub fn new(
 		mempool: T,
 		block_size: u32,
-		parent_block: Arc<RwLock<Id>>,
+		parent_block: Arc<RwLock<block::Id>>,
 		building_time_ms: u64,
 	) -> Self {
 		Self { mempool, block_size, parent_block, building_time_ms }
@@ -50,7 +53,7 @@ impl Memseq<RocksdbMempool> {
 		let mempool = RocksdbMempool::try_new(
 			path.to_str().ok_or(anyhow::anyhow!("PathBuf to str failed"))?,
 		)?;
-		let parent_block = Arc::new(RwLock::new(Id::default()));
+		let parent_block = Arc::new(RwLock::new(block::Id::default()));
 		Ok(Self::new(mempool, block_size, parent_block, building_time_ms))
 	}
 
@@ -153,7 +156,7 @@ pub mod test {
 	#[tokio::test]
 	async fn test_publish_error_propagation() -> Result<(), anyhow::Error> {
 		let mempool = MockMempool;
-		let parent_block = Arc::new(RwLock::new(Id::default()));
+		let parent_block = Arc::new(RwLock::new(block::Id::default()));
 		let memseq = Memseq::new(mempool, 10, parent_block, 1000);
 
 		let transaction = Transaction::new(vec![1, 2, 3], 0);
@@ -245,7 +248,7 @@ pub mod test {
 		)?;
 		let block_size = 50;
 		let building_time_ms = 2000;
-		let parent_block = Arc::new(RwLock::new(Id::default()));
+		let parent_block = Arc::new(RwLock::new(block::Id::default()));
 
 		let memseq = Memseq::new(mem_pool, block_size, Arc::clone(&parent_block), building_time_ms);
 
@@ -266,7 +269,7 @@ pub mod test {
 		)?;
 		let block_size = 50;
 		let building_time_ms = 2000;
-		let parent_block = Arc::new(RwLock::new(Id::default()));
+		let parent_block = Arc::new(RwLock::new(block::Id::default()));
 
 		let memseq = Memseq::new(mem_pool, block_size, Arc::clone(&parent_block), building_time_ms);
 
@@ -413,7 +416,7 @@ pub mod test {
 	impl MempoolTransactionOperations for MockMempool {
 		async fn has_mempool_transaction(
 			&self,
-			_transaction_id: Id,
+			_transaction_id: transaction::Id,
 		) -> Result<bool, anyhow::Error> {
 			Err(anyhow::anyhow!("Mock has_mempool_transaction"))
 		}
@@ -434,7 +437,7 @@ pub mod test {
 
 		async fn remove_mempool_transaction(
 			&self,
-			_transaction_id: Id,
+			_transaction_id: transaction::Id,
 		) -> Result<(), anyhow::Error> {
 			Err(anyhow::anyhow!("Mock remove_mempool_transaction"))
 		}
@@ -447,7 +450,7 @@ pub mod test {
 
 		async fn get_mempool_transaction(
 			&self,
-			_transaction_id: Id,
+			_transaction_id: transaction::Id,
 		) -> Result<Option<MempoolTransaction>, anyhow::Error> {
 			Err(anyhow::anyhow!("Mock get_mempool_transaction"))
 		}
@@ -462,7 +465,7 @@ pub mod test {
 	}
 
 	impl MempoolBlockOperations for MockMempool {
-		async fn has_block(&self, _block_id: Id) -> Result<bool, anyhow::Error> {
+		async fn has_block(&self, _block_id: block::Id) -> Result<bool, anyhow::Error> {
 			todo!()
 		}
 
@@ -470,11 +473,11 @@ pub mod test {
 			todo!()
 		}
 
-		async fn remove_block(&self, _block_id: Id) -> Result<(), anyhow::Error> {
+		async fn remove_block(&self, _block_id: block::Id) -> Result<(), anyhow::Error> {
 			todo!()
 		}
 
-		async fn get_block(&self, _block_id: Id) -> Result<Option<Block>, anyhow::Error> {
+		async fn get_block(&self, _block_id: block::Id) -> Result<Option<Block>, anyhow::Error> {
 			todo!()
 		}
 	}
