+++ protocol-units/settlement/mcr/client/src/eth_client.rs
@@ -21,8 +21,7 @@ use alloy_transport::BoxTransport;
 use alloy_transport_ws::WsConnect;
 use anyhow::Context;
 use mcr_settlement_config::Config;
-use movement_types::BlockCommitment;
-use movement_types::{Commitment, Id};
+use movement_types::block::{BlockCommitment, Commitment, Id};
 use serde_json::Value as JsonValue;
 use std::array::TryFromSliceError;
 use std::fs;
